Legacy screens cleanup New fragment or showing EFGS update created Unused resources cleanup
