--- app/src/main/kotlin/cz/covid19cz/erouska/ui/update/legacy/LegacyUpdateVM.kt
@@ -1,40 +1,0 @@
-package cz.covid19cz.erouska.ui.update.legacy
-
-import androidx.hilt.lifecycle.ViewModelInject
-import androidx.lifecycle.MutableLiveData
-import arch.viewmodel.BaseArchViewModel
-import cz.covid19cz.erouska.db.SharedPrefsRepository
-import cz.covid19cz.erouska.ui.update.legacy.event.LegacyUpdateEvent
-
-class LegacyUpdateVM @ViewModelInject constructor(
-     val sharedPrefsRepository: SharedPrefsRepository
-) : BaseArchViewModel() {
-
-    val state = MutableLiveData<LegacyUpdateEvent>()
-
-    fun next() {
-        val next = when (state.value) {
-            LegacyUpdateEvent.LegacyUpdateExpansion -> LegacyUpdateEvent.LegacyUpdateActiveNotification
-            LegacyUpdateEvent.LegacyUpdateActiveNotification -> LegacyUpdateEvent.LegacyUpdatePhoneNumbers
-            LegacyUpdateEvent.LegacyUpdatePhoneNumbers -> LegacyUpdateEvent.LegacyUpdatePrivacy
-            else -> LegacyUpdateEvent.LegacyUpdateFinish
-        }
-
-        state.value = next
-    }
-
-    fun previous() {
-        val prev = when (state.value) {
-            LegacyUpdateEvent.LegacyUpdatePrivacy -> LegacyUpdateEvent.LegacyUpdatePhoneNumbers
-            LegacyUpdateEvent.LegacyUpdatePhoneNumbers -> LegacyUpdateEvent.LegacyUpdateActiveNotification
-            else -> LegacyUpdateEvent.LegacyUpdateExpansion
-        }
-
-        state.value = prev
-    }
-
-    fun finish() {
-        sharedPrefsRepository.markUpdateFromLegacyVersionCompleted()
-    }
-
-}
