--- app/src/main/kotlin/cz/covid19cz/erouska/ui/update/legacy/event/LegacyUpdateCommandEvent.kt
@@ -1,10 +1,0 @@
-package cz.covid19cz.erouska.ui.update.legacy.event
-
-sealed class LegacyUpdateEvent {
-    object LegacyUpdateEFGS : LegacyUpdateEvent()
-    object LegacyUpdateExpansion : LegacyUpdateEvent()
-    object LegacyUpdateActiveNotification : LegacyUpdateEvent()
-    object LegacyUpdatePhoneNumbers : LegacyUpdateEvent()
-    object LegacyUpdatePrivacy : LegacyUpdateEvent()
-    object LegacyUpdateFinish : LegacyUpdateEvent()
-}
