+++ app/src/main/kotlin/cz/covid19cz/erouska/ui/welcome/WelcomeFragment.kt
@@ -30,10 +30,6 @@ class WelcomeFragment :
         super.onViewCreated(view, savedInstanceState)
         enableUpInToolbar(false)
 
-        if (viewModel.wasAppUpdated()) {
-            navigate(R.id.action_nav_welcome_fragment_to_nav_legacy_update)
-        }
-
         val welcomeDescription: String = String.format(
             getString(R.string.welcome_description)
         )
