--- app/src/main/res/drawable/ic_update_phone.xml
@@ -1,73 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="585.45dp"
-    android:height="582.47dp"
-    android:viewportWidth="585.45"
-    android:viewportHeight="582.47">
-  <path
-      android:pathData="M290.67,291.81m-262.51,0a262.51,262.51 0,1 1,525.02 0a262.51,262.51 0,1 1,-525.02 0"
-      android:fillColor="#cae2f4"/>
-  <path
-      android:pathData="M292,2.15C132.12,2.15 2.48,131.8 2.48,291.72S132.12,581.28 292,581.28 581.6,451.64 581.6,291.72 452,2.15 292,2.15ZM292,553C147.75,553 30.78,436 30.78,291.72S147.75,30.46 292,30.46s261.26,117 261.26,261.26S436.33,553 292,553Z"
-      android:fillColor="#e5f2f9"/>
-  <path
-      android:pathData="M223.327,112.687L366.947,112.511A4.27,4.27 134.93,0 1,371.222 116.776L371.596,422.126A4.27,4.27 134.93,0 1,367.331 426.401L223.711,426.577A4.27,4.27 134.93,0 1,219.436 422.312L219.063,116.962A4.27,4.27 134.93,0 1,223.327 112.687z"
-      android:fillColor="#334176"/>
-  <path
-      android:pathData="M219.071,131.872l152.16,-0.186l0.326,266.53l-152.16,0.186z"
-      android:fillColor="#fefeff"/>
-  <path
-      android:pathData="M219.146,184.862l152.16,-0.186l0.21,172.06l-152.16,0.186z"
-      android:fillColor="#c4cae2"/>
-  <path
-      android:pathData="M247,207h99.6c2.45,1.32 2.95,3.45 2.95,6.08q-0.09,54.66 0,109.32c0,5 -1.2,6.12 -6.26,6.12 -30.66,0 -61.32,-0.11 -92,0.09 -6.79,0.05 -12.22,-3.43 -12.15,-12.22 0.25,-32.3 0.19,-64.6 0,-96.9C239.12,213.31 241.11,209.09 247,207ZM324,215.48L251.36,215.48c-3.78,0 -3.78,0 -3.78,3.7v97.18c0,3.68 0,3.69 3.8,3.69L324,320.05ZM332.69,244.48L341,244.48v-28.9h-8.32ZM332.69,253.07L332.69,282L341,282L341,253.1ZM341,319.9L341,291.05h-8.3L332.7,319.9Z"
-      android:fillColor="#334176"/>
-  <path
-      android:pathData="M285.08,257.83l9.24,15.93c3.67,-0.14 6.71,-0.29 9.75,-0.33a2.73,2.73 0,0 1,1.72 0.71c5,4.71 8.89,10.18 9.35,17.21a8.5,8.5 0,0 1,-2.65 6.15c-6.23,5.22 -13.21,8.73 -21.8,6.77a20.23,20.23 0,0 1,-13.38 -9.7q-9.37,-15.9 -18.48,-32a20.61,20.61 0,0 1,6.59 -27,65.61 65.61,0 0,1 6.26,-3.71 9.77,9.77 0,0 1,13.22 3.52c1.59,2.5 3,5.11 4.46,7.7 2.94,5.29 1.87,10.11 -3,13.72ZM264.37,252.92c0.81,2.24 1.33,4.64 2.48,6.68 5.6,9.95 11.33,19.81 17.07,29.67 4.31,7.38 11.17,9.18 18.65,4.87 5.58,-3.23 5.35,-2.1 2,-7.82 -3.15,-5.44 -3.16,-5.43 -8.59,-2.51 -3,1.6 -4.94,1.15 -6.59,-1.67q-6.81,-11.62 -13.51,-23.31c-1.61,-2.79 -1,-4.89 1.78,-6.58 5.34,-3.27 5.37,-3.3 2.25,-8.58 -3.5,-5.92 -2.18,-5.47 -8.27,-2.06A13.64,13.64 0,0 0,264.37 252.92Z"
-      android:fillColor="#334176"/>
-  <group>
-    <clip-path
-        android:pathData="M292.56,290.59m-261.26,0a261.26,261.26 0,1 1,522.52 0a261.26,261.26 0,1 1,-522.52 0"/>
-    <path
-        android:pathData="M270.47,633.4c-40.06,0.11 -80.17,-1 -120.15,0.81 -21.29,1 -20.63,-6.58 -19.94,-26.42 0.16,-4.31 8.93,-31.94 11.44,-42.95 5.8,-15.72 10.3,-31.88 16,-47.6 3.26,-8.9 6.22,-18.1 16.08,-22.3 41.45,-5.54 79,2.49 110.43,31.85 9.05,22.33 -0.44,44.26 -1.46,66.38C281.8,607.52 282.42,622.41 270.47,633.4Z"
-        android:fillColor="#ba417e"/>
-    <path
-        android:pathData="M172.75,498.8c-10.85,-23.8 9.84,-20.12 21.3,-24.49 29.13,-6.74 55.54,-2.95 77.05,19.89 16.59,7 20.54,17.6 9.3,32.55C246.83,508.47 211.44,497.25 172.75,498.8Z"
-        android:fillColor="#2d2f62"/>
-    <path
-        android:pathData="M271.1,494.2A265.83,265.83 0,0 0,194 474.31c1.92,-23.91 3.29,-47.31 0,-70.69 -6.52,-47 0.83,-74.18 12.85,-118.92 0.44,-1.64 6.94,-15.85 8,-17.31 2.14,-2.81 4.83,-5.35 4.64,1.26 -0.25,9.08 0.49,9.78 0.49,19.76 0,19 -1.59,40.19 1,59.11a93.3,93.3 0,0 0,8.62 -3.9,15.21 15.21,0 0,0 2.8,-3.55c9.83,-13.14 18.28,-27.21 29.83,-38.82 6.33,-6.37 13.44,-13 23.31,-8.39 9.11,4.25 1.67,14.87 -2.46,23.13 -6.06,12.09 -13.18,24.2 -20.41,36.51 -5.72,9.74 -10,16.33 -12.38,34 -4.94,17.84 0.77,26.09 0.33,39.71 10.43,0.47 45.72,0.25 49.57,0.25C297.91,451.76 279.42,471.61 271.1,494.2Z"
-        android:fillColor="#e1a16e"/>
-  </group>
-  <path
-      android:pathData="M357.714,203.713L357.714,203.713A4.92,4.915 134.998,0 1,357.711 210.668L237.46,330.918A4.92,4.915 134.998,0 1,230.506 330.922L230.506,330.922A4.92,4.915 134.998,0 1,230.509 323.967L350.76,203.717A4.92,4.915 134.998,0 1,357.714 203.713z"
-      android:fillColor="#ae005a"/>
-  <path
-      android:pathData="M357.707,330.934L357.707,330.934A4.915,4.92 134.998,0 1,350.752 330.931L230.502,210.68A4.915,4.92 134.998,0 1,230.498 203.726L230.498,203.726A4.915,4.92 134.998,0 1,237.453 203.729L357.703,323.98A4.915,4.92 134.998,0 1,357.707 330.934z"
-      android:fillColor="#ae005a"/>
-  <path
-      android:pathData="M245.58,159.87m-16.73,0a16.73,16.73 0,1 1,33.46 0a16.73,16.73 0,1 1,-33.46 0"
-      android:fillColor="#bd4b85"/>
-  <path
-      android:pathData="M245.66,154.06m-4.28,0a4.28,4.28 0,1 1,8.56 0a4.28,4.28 0,1 1,-8.56 0"
-      android:fillColor="#fefeff"/>
-  <path
-      android:pathData="M252.45,167.36c0,0.67 -3.49,1.78 -7.11,1.74s-7.36,-1.24 -7.36,-1.74a7.24,7.24 0,1 1,14.47 0Z"
-      android:fillColor="#fefeff"/>
-  <path
-      android:pathData="M269.14,144.95h92.04v29.96h-92.04z"
-      android:fillColor="#e9edf8"/>
-  <path
-      android:pathData="M272.62,149.24h85.15v4.74h-85.15z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M272.78,158.07h36.33v4.74h-36.33z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M325.91,158.28h31.41v4.74h-31.41z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M272.93,166.3h42.01v4.74h-42.01z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M326.06,166.4h31.41v4.74h-31.41z"
-      android:fillColor="#9c8cad"/>
-</vector>
