--- app/src/main/res/drawable/ic_update_privacy.xml
@@ -1,85 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="585.45dp"
-    android:height="582.47dp"
-    android:viewportWidth="585.45"
-    android:viewportHeight="582.47">
-  <path
-      android:pathData="M290.67,291.81m-262.51,0a262.51,262.51 0,1 1,525.02 0a262.51,262.51 0,1 1,-525.02 0"
-      android:fillColor="#cae2f4"/>
-  <path
-      android:pathData="M292,2.15C132.12,2.15 2.48,131.8 2.48,291.72S132.12,581.28 292,581.28 581.6,451.64 581.6,291.72 452,2.15 292,2.15ZM292,553C147.75,553 30.78,436 30.78,291.72S147.75,30.46 292,30.46s261.26,117 261.26,261.26S436.33,553 292,553Z"
-      android:fillColor="#e5f2f9"/>
-  <path
-      android:pathData="M225.298,109.903L369.787,110.155A4.27,4.27 45.1,0 1,374.05 114.432L373.514,421.562A4.27,4.27 45.1,0 1,369.236 425.824L224.747,425.572A4.27,4.27 45.1,0 1,220.484 421.295L221.02,114.165A4.27,4.27 45.1,0 1,225.298 109.903z"
-      android:fillColor="#334176"/>
-  <path
-      android:pathData="M220.994,129.175l153.03,0.267l-0.468,268.04l-153.03,-0.267z"
-      android:fillColor="#fefeff"/>
-  <path
-      android:pathData="M220.828,224.165l153.03,0.267l-0.229,131.34l-153.03,-0.267z"
-      android:fillColor="#c4cae2"/>
-  <path
-      android:pathData="M255.84,249.8m-11.92,0a11.92,11.92 0,1 1,23.84 0a11.92,11.92 0,1 1,-23.84 0"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M237.783,280.737l115.659,0.545l-0.169,35.77l-115.659,-0.545z"
-      android:fillColor="#fefefe"/>
-  <path
-      android:pathData="M282.106,237.747l69.159,0.326l-0.022,4.77l-69.159,-0.326z"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M282.068,247.846l69.159,0.326l-0.022,4.77l-69.159,-0.326z"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M236.917,330.903l115.819,0.546l-0.022,4.77l-115.819,-0.546z"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M236.873,340.193l115.819,0.546l-0.022,4.77l-115.819,-0.546z"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M283.24,257.5h69.16v4.77h-69.16z"
-      android:fillColor="#a1aacd"/>
-  <path
-      android:pathData="M227.111,153.345l139.93,0.317l-0.103,45.55l-139.93,-0.317z"
-      android:fillColor="#e9edf8"/>
-  <path
-      android:pathData="M283.93,159.985l77.93,0.136l-0.013,7.21l-77.93,-0.136z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M313.386,173.797l47.76,0.083l-0.013,7.21l-47.76,-0.083z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M313.575,186.127l47.76,0.083l-0.013,7.21l-47.76,-0.083z"
-      android:fillColor="#9c8cad"/>
-  <path
-      android:pathData="M251.28,176.9m-24.15,0a24.15,24.15 0,1 1,48.3 0a24.15,24.15 0,1 1,-48.3 0"
-      android:fillColor="#bd4b85"/>
-  <path
-      android:pathData="M251.4,168.51m-6.18,0a6.18,6.18 0,1 1,12.36 0a6.18,6.18 0,1 1,-12.36 0"
-      android:fillColor="#fefeff"/>
-  <path
-      android:pathData="M261.2,187.7c0,1 -5,2.58 -10.26,2.52s-10.62,-1.79 -10.62,-2.52a10.44,10.44 0,0 1,20.88 0Z"
-      android:fillColor="#fefeff"/>
-  <group>
-    <clip-path
-        android:pathData="M292.56,290.59m-261.26,0a261.26,261.26 0,1 1,522.52 0a261.26,261.26 0,1 1,-522.52 0"/>
-    <path
-        android:pathData="M270.47,633.4c-40.06,0.11 -80.17,-1 -120.15,0.81 -21.29,1 -20.63,-6.58 -19.94,-26.42 0.16,-4.31 8.93,-31.94 11.44,-42.95 5.8,-15.72 10.3,-31.88 16,-47.6 3.26,-8.9 6.22,-18.1 16.08,-22.3 41.45,-5.54 79,2.49 110.43,31.85 9.05,22.33 -0.44,44.26 -1.46,66.38C281.8,607.52 282.42,622.41 270.47,633.4Z"
-        android:fillColor="#ba417e"/>
-    <path
-        android:pathData="M172.75,498.8c-10.85,-23.8 9.84,-20.12 21.3,-24.49 29.13,-6.74 55.54,-2.95 77.05,19.89 16.59,7 20.54,17.6 9.3,32.55C246.83,508.47 211.44,497.25 172.75,498.8Z"
-        android:fillColor="#2d2f62"/>
-    <path
-        android:pathData="M271.1,494.2A265.83,265.83 0,0 0,194 474.31c1.92,-23.91 3.29,-47.31 0,-70.69 -6.52,-47 0.83,-74.18 12.85,-118.92 0.44,-1.64 6.94,-15.85 8,-17.31 2.14,-2.81 4.83,-5.35 4.64,1.26 -0.25,9.08 0.49,9.78 0.49,19.76 0,19 -1.59,40.19 1,59.11a93.3,93.3 0,0 0,8.62 -3.9,15.21 15.21,0 0,0 2.8,-3.55c9.83,-13.14 18.28,-27.21 29.83,-38.82 6.33,-6.37 13.44,-13 23.31,-8.39 9.11,4.25 1.67,14.87 -2.46,23.13 -6.06,12.09 -13.18,24.2 -20.41,36.51 -5.72,9.74 -10,16.33 -12.38,34 -4.94,17.84 0.77,26.09 0.33,39.71 10.43,0.47 45.72,0.25 49.57,0.25C297.91,451.76 279.42,471.61 271.1,494.2Z"
-        android:fillColor="#e1a16e"/>
-  </group>
-  <path
-      android:pathData="M352.39,78.06A73.91,73.91 0,0 0,278.48 152c0,40.82 33.09,42.34 73.91,42.34S426.3,192.79 426.3,152A73.91,73.91 0,0 0,352.39 78.06ZM352.39,168.98c-26.69,0 -48.32,7.5 -48.32,-19.19a48.33,48.33 0,0 1,96.65 0C400.72,176.48 379.08,169 352.39,169Z"
-      android:fillColor="#1c2244"/>
-  <path
-      android:pathData="M284.57,162.72h134.92v114.09h-134.92z"
-      android:fillColor="#fff"/>
-  <path
-      android:pathData="M262.11,162.37c32,-0.2 60.49,-8.71 84.47,-29.54a8.84,8.84 0,0 1,11.59 0,116.85 116.85,0 0,0 36.48,21.59c15.49,5.63 31.54,7.8 48,8a1.79,1.79 0,0 1,1.78 1.77h0c0.14,15.78 0.56,3.51 0.32,19.29 -0.28,17.9 -1.84,37.45 -6.54,54.85 -8.54,31.61 -26.85,55.71 -55.17,72.21A175.42,175.42 0,0 1,353.6 324a4.07,4.07 0,0 1,-2.32 0c-15.86,-5.63 -30.78,-13 -44.12,-23.4 -18.58,-14.49 -31.24,-33.06 -38.5,-55.39a162.19,162.19 0,0 1,-7.57 -38.2c-1.47,-17.36 -1.25,-10.17 -0.9,-27.56 0.11,-5 0,-10 0,-15.2A1.9,1.9 0,0 1,262.11 162.37ZM346.6,263.58l57.45,-57.26a3.3,3.3 0,0 0,0 -4.68L389.46,187a3.38,3.38 0,0 0,-4.78 0l-41,41.23a3.38,3.38 0,0 1,-4.8 0L320,208.88a3.33,3.33 0,0 0,-4.72 0l-14.88,14.77a3.31,3.31 0,0 0,0 4.69l35.16,35.24A7.82,7.82 0,0 0,346.6 263.58Z"
-      android:fillColor="#28325c"/>
-</vector>
