+++ app/src/main/res/navigation/nav_graph.xml
@@ -5,18 +5,26 @@
     android:id="@+id/nav_graph">
 
     <fragment
-        android:id="@+id/nav_legacy_update_fragment"
-        android:name="cz.covid19cz.erouska.ui.update.legacy.LegacyUpdateFragment"
-        android:label="@string/legacy_update_title"
-        tools:layout="@layout/fragment_legacy_update">
+        android:id="@+id/nav_efgs_update_fragment"
+        android:name="cz.covid19cz.erouska.ui.update.efgs.EfgsUpdateFragment"
+        android:label="@string/efgs_title"
+        tools:layout="@layout/fragment_efgs_update">
         <argument
             android:name="fullscreen"
             android:defaultValue="false"
             app:argType="boolean" />
         <argument
-            android:name="efgs"
+            android:name="onboarding"
             android:defaultValue="false"
             app:argType="boolean" />
+
+        <action
+            android:id="@+id/action_nav_efgs_update_fragment_to_activation_fragment"
+            app:destination="@+id/nav_activation"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
     </fragment>
 
     <fragment
@@ -28,6 +36,14 @@
         <action
             android:id="@+id/action_nav_activation_notifications_to_activation_fragment"
             app:destination="@+id/nav_activation"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+
+        <action
+            android:id="@+id/action_nav_activation_notifications_to_efgs_update"
+            app:destination="@+id/nav_efgs_update_fragment"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
@@ -53,8 +69,8 @@
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
             app:popExitAnim="@anim/nav_default_pop_exit_anim" />
         <action
-            android:id="@+id/action_nav_welcome_fragment_to_nav_legacy_update"
-            app:destination="@+id/nav_legacy_update_fragment"
+            android:id="@+id/action_nav_welcome_fragment_to_nav_efgs_update"
+            app:destination="@+id/nav_efgs_update_fragment"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
@@ -234,8 +250,8 @@
         android:label="@string/app_name"
         tools:layout="@layout/fragment_dashboard_cards">
         <action
-            android:id="@+id/action_nav_dashboard_to_nav_legacy_update"
-            app:destination="@id/nav_legacy_update_fragment"
+            android:id="@+id/action_nav_dashboard_to_nav_efgs_update"
+            app:destination="@id/nav_efgs_update_fragment"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
